requirements.txt
YYYTools>=0.1.1
django>=3.0.3
matplotlib>=2.2.3
pandas>=0.25.1